Split Contents and Sections into separate controllers
